fix(make check): Use of development verion of gnunet-service-mqtt for make check
Testbed was configured to use the installed version of gnunet-service-mqtt
located in /usr/local for testing when make check was run. This had the big
disadvantage that in order to test changes in gnunet-service-mqtt one had to do
a make install first. This fix is now changed.

What this change does is the following. The configuration of the tests has been
changed to explicitly tell testbed which binary to use. For this to be platform
independent the config file will be parsed by a small perl script to resolve any
${VAR} occurences in the configuration file with the bash environmental
variable.

Closes #6